requirements.txt
click==8.1.3
loguru==0.7.0
python-daemon==3.0.1
web3==6.5.0
